# Conflicts:
#	src/training/pango/boxchar.cpp
#	src/training/pango/stringrenderer.cpp

+++ src/ccmain/par_control.cpp
@@ -61,7 +61,7 @@ void Tesseract::PrerecAllWordsPar(const std::vector<WordData> &words) {
 #endif // _OPENMP
     // NOLINTNEXTLINE(modernize-loop-convert)
 	// Error C3016: index variable in OpenMP 'for' statement must have signed integral type
-    for (int b = 0; b < static_cast<int>(blobs.size()); ++b) {
+    for (size_t b = 0; b < blobs.size(); ++b) {
       *blobs[b].choices =
           blobs[b].tesseract->classify_blob(blobs[b].blob, "par", ScrollView::WHITE, nullptr);
     }
